pyproject.toml
[tool.poetry]
name = "TournamentStreamHelper"
version = "6.0.0"
description = "Tournament scoreboard helper and overlays for fighting game tournaments"
authors = ["João Ribeiro Bezerra <joaorb64@gmail.com>", "Isatis <isatis@foxes.gay>", "Mathias Wolfbrok <wolfy76700>", "Eion Dailey <zombiehdgaming>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
httpx = "^0.27.0"
deepdiff = "^7.0.1"
asyncio = "^3.4.3"
python-dateutil = "^2.9.0.post0"
loguru = "^0.7.2"
orjson = "^3.10.3"
websockets = "^12.0"
python-socketio = "^5.11.2"
aiofiles = "^23.2.1"
hypercorn = "^0.17.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
